requirements.txt
python-binance
pandas
ccxt
fastapi
uvicorn
matplotlib
flask
dash
plotly
tradingpattern
colorama
psycopg2-binary
psycopg2
websocket-client
ta